fix: pushed connect to latest view component
